移动集成化的程序
