+++ TeamCode/src/main/java/org/firstinspires/ftc/opmodes/autonomous/LeftStale.java
@@ -13,7 +13,7 @@ import com.acmerobotics.roadrunner.trajectory.Trajectory;
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
 
 import org.acmerobotics.roadrunner.trajectorysequence.TrajectorySequence;
-import org.firstinspires.ftc.teamcode.cores.eventloop.StructuralLinearMode;
+import org.firstinspires.ftc.teamcode.cores.eventloop.integral.StructuralLinearMode;
 import org.firstinspires.ftc.teamcode.cores.structure.SimpleDriveOp;
 
 @Config
