+++ TeamCode/src/main/java/org/firstinspires/ftc/opmodes/autonomous/RightStale.java
@@ -10,7 +10,7 @@ import com.acmerobotics.dashboard.config.Config;
 import com.acmerobotics.roadrunner.geometry.Pose2d;
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
 
-import org.firstinspires.ftc.teamcode.cores.eventloop.IntegralAutonomous;
+import org.firstinspires.ftc.teamcode.cores.eventloop.integral.IntegralAutonomous;
 import org.firstinspires.ftc.teamcode.cores.structure.SimpleDriveOp;
 
 @Config
