+++ TeamCode/src/main/java/org/firstinspires/ftc/opmodes/tests/TestAutonomous.java
@@ -3,7 +3,7 @@ package org.firstinspires.ftc.opmodes.tests;
 import com.acmerobotics.dashboard.config.Config;
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
 
-import org.firstinspires.ftc.teamcode.cores.eventloop.IntegralAutonomous;
+import org.firstinspires.ftc.teamcode.cores.eventloop.integral.IntegralAutonomous;
 
 @Autonomous(group = "9_Tests")
 @Config
