+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/cores/eventloop/integral/IntegralOpMode.java
@@ -1,4 +1,7 @@
-package org.firstinspires.ftc.teamcode.cores.eventloop;
+package org.firstinspires.ftc.teamcode.cores.eventloop.integral;
+
+import org.firstinspires.ftc.teamcode.cores.eventloop.OpTerminateException;
+import org.firstinspires.ftc.teamcode.cores.eventloop.TerminateReason;
 
 public interface IntegralOpMode {
 	default void sendTerminateSignal(final TerminateReason reason){
