+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/cores/eventloop/integral/StructuralLinearMode.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.cores.eventloop;
+package org.firstinspires.ftc.teamcode.cores.eventloop.integral;
 
 /**
  * 提供结构化方法
