+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/events/IntegralThreadExceptionHandler.java
@@ -9,8 +9,8 @@ import androidx.annotation.NonNull;
 import com.qualcomm.robotcore.eventloop.opmode.OpModeManagerImpl;
 
 import org.betastudio.ftc.Interfaces;
-import org.firstinspires.ftc.teamcode.cores.eventloop.IntegralAutonomous;
-import org.firstinspires.ftc.teamcode.cores.eventloop.IntegralOpMode;
+import org.firstinspires.ftc.teamcode.cores.eventloop.integral.IntegralAutonomous;
+import org.firstinspires.ftc.teamcode.cores.eventloop.integral.IntegralOpMode;
 import org.firstinspires.ftc.teamcode.cores.eventloop.TerminateReason;
 
 import java.lang.Thread.UncaughtExceptionHandler;
